ai-reporting/requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
requests>=2.31
jinja2>=3.1
